backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
bidict==0.21.4
charset-normalizer==2.0.12
click==8.0.4
cmake==3.22.2
colorama==0.4.4
cycler==0.11.0
dlib==19.23.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-SocketIO==4.3.1
fonttools==4.29.1
frozenlist==1.3.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.1
imutils==0.5.4
itsdangerous==2.1.0
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.1.0
matplotlib==3.5.1
multidict==6.0.2
numpy==1.21.5
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
python-engineio==3.13.2
python-socketio==4.6.0
six==1.16.0
typing_extensions==4.1.1
Werkzeug==2.0.3
yarl==1.7.2
zipp==3.7.0